requirements.txt
pytelegrambotapi~=4.6.1
requests~=2.28.1
urllib3~=1.26.11
rich~=12.5.1
PyYAML~=6.0